Fixing parsing empty string error from client to server
In nhucauthucte component: Initialize variable "quantity" to 0 instead of an empty string,
before sending data to server parse variable to isNaN(parseInt(quantity)) ? 0 : parseInt(quantity)
func to make sure the typo is integer, if NaN then make value to 0.

In capnhatnctt component: Initialize variable ait and kmt to 0 and before sending data to server,
check if user input empty string or not, if NaN then make it to 0, else parse it to parseInt func
to make it an integer

In nhucauthucte.routes.js: Remove ait and kmt from post because the default value in the database of
those 2 variables will be set to 0.